requirements.txt
Django==2.1.14
django-cors-headers==3.1.1
django-filter==2.2.0
django-mobile==0.7.0
django-mysql==3.2.0
django-rest-knox==4.1.0
django-restframework==0.0.1
djangorestframework==3.10.3
MarkupSafe==1.1.1
mysqlclient==1.4.4
PyMySQL==0.9.3
pyOpenSSL==17.0.0
sqlparse==0.3.0
virtualenv==16.0.0
webassets==0.12.1
webencodings==0.5
websocket-client==0.53.0
Werkzeug==0.15.6